Added a close register modal, added TODOs for fireabse js, app.jsx,eventsignupform.jsx, eventsData.jsx. Added eventData which needs to be dynamic
