+++ WA-script/PyWhatKit_DB.txt
@@ -193,3 +193,8 @@ Time: 14:39
 Phone Number: +6281239654795
 Message: Hello Enrico Pratama :-)Thank you for signing up for ISAUW's Seattle Stamp Quest. We're excited to meet you and here are some reminders for today:1. Meet us at Red Square (in front of Suzzallo) at 2pm SHARP.2. Bring your ORCA or UW ID card as we will be taking public transport throughout the day.Attached below are your assigned groups! If you have any questions or concerns or if you are unable to come to the event, please feel free to DM us on Instagram @isauwhuskies . Thank you so much for supporting us through this event and we hope to see you there!https://docs.google.com/document/d/1pHejMVhsmmP-OeQUpXwoquhl2bxWYuxpR6Rm_tfSxNc/
 --------------------
+Date: 27/9/2024
+Time: 11:40
+Phone Number: +6281239654795
+Message: Hello Enrico Pratama :-)Thank you for signing up for ISAUW's Seattle Stamp Quest. We're excited to meet you and here are some reminders for today:1. Meet us at Red Square (in front of Suzzallo) at 2pm SHARP.2. Bring your ORCA or UW ID card as we will be taking public transport throughout the day.Attached below are your assigned groups! If you have any questions or concerns or if you are unable to come to the event, please feel free to DM us on Instagram @isauwhuskies . Thank you so much for supporting us through this event and we hope to see you there!https://docs.google.com/document/d/1pHejMVhsmmP-OeQUpXwoquhl2bxWYuxpR6Rm_tfSxNc/
+--------------------
