+++ src/components/App.jsx
@@ -14,17 +14,36 @@ import ScrollToTop from "./ScrollToTop";
 import SocialLinks from "./SocialLinks";
 import Apply from "./Apply/Apply";
 import "./index.css"; // Tailwind CSS
-import SignUpFormComponent from "./Events/SignUpFormComp";
+import EventSignUpForm from "./Events/EventSignUpForm";
+import eventsData from "../data/eventsData";
 
 function App() {
   const [expanded, setExpanded] = useState(false);
   const [navBar, setNavBar] = useState(false);
   const [keratonPage, setKeratonPage] = useState(false);
   const [keratonScrollDown, setKeratonScrollDown] = useState(false);
+  const [events, setEvents] = useState([]);
+  const [uncompletedEvents, setUncompletedEvents] = useState([]);
   const [currentPage, setCurrentPage] = useState(
     window.location.pathname.substring(1)
   );
   const [isMobile, setIsMobile] = useState(false);
+
+  const getEvents = () => {
+    const events_ = [];
+    const uncompletedEvents_ = [];
+    for (const event of eventsData) {
+      if (!event.completed) {
+        uncompletedEvents_.push(event);
+      }
+      events_.push(event);
+    }
+
+    setEvents(events_);
+    setUncompletedEvents(uncompletedEvents_);
+  };
+
+  useEffect(getEvents, []);
 
   const resetHeight = () => {
     // reset the body height to that of the inner browser
@@ -231,11 +250,41 @@ function App() {
         <Switch>
           <Route path="/" exact render={() => <Home isMobile={isMobile} />} />
           <Route path="/events" component={Events} />
+          {/* TODO: make this auto generate for each of eventsData.js.eventsData.completed === false  */}
+
+          {/* Dynamically generate routes for uncompleted events */}
+          {/* {uncompletedEvents.map((uncompletedEvent, index) => {
+            const eventPath = `/sign-up-${uncompletedEvent.title
+              .toLowerCase()
+              .replace(/ /g, "-")}`;
+
+            return (
+              <Route
+                key={index}
+                path={eventPath}
+                render={() => (
+                  <EventSignUpForm
+                    eventName={uncompletedEvent.title}
+                    BGImage={
+                      uncompletedEvent.BGImage || "/default/background.png"
+                    } // Default background if none provided
+                    firestorePath={
+                      uncompletedEvent.firestorePath || "default/firestore/path"
+                    } // Default path if none provided
+                    posterImage={
+                      uncompletedEvent.img[0]?.src || "/default/poster.png"
+                    } // Default poster if none provided
+                    rsvp={uncompletedEvent.rsvp || false} // Default RSVP status
+                  />
+                )}
+              />
+            );
+          })} */}
           <Route
-            path="/sign-up-stamp-quest"
+            path="/sign-up-test-event"
             render={() => (
-              <SignUpFormComponent
-                eventName="Seattle Stamp Quest"
+              <EventSignUpForm
+                eventName="Test Event"
                 BGImage="/events/stampquest/stampQuestFormBG.png"
                 firestorePath="2024/stamp-quest/event-registrations"
                 posterImage="/events/stampquest/stamp_quest_poster.png"
@@ -245,7 +294,7 @@ function App() {
           <Route
             path="/sign-up-seathrough"
             render={() => (
-              <SignUpFormComponent
+              <EventSignUpForm
                 eventName="Seathrough"
                 BGImage="/events/seathrough/2024/seathrough_banner.png"
                 firestorePath="2024/seathrough/event-registrations"
